pyproject.toml
[project]
name = "medfm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch==2.9.1",
    "torchvision",
    "torchaudio",
    "timm>=1.0.17",
    "numpy==2.1.2",
    "pillow==11.0.0",
    "pandas==2.2.3",
    "scikit-learn==1.6.1",
    "matplotlib==3.10.1",
    "seaborn==0.13.2",
    "medpy==0.4.0",
    "termcolor==3.0.1",
    "tensorboardx==2.6.2.2",
    "scikit-image>=0.25.2",
    "hydra-core>=1.3.2",
    "tqdm>=4.67.1",
    "wandb>=0.23.0",
    "icecream>=2.1.8",
    "einops>=0.8.1",
    "safetensors>=0.7.0",
    "transformers>=4.57.3",
    "python-dotenv>=1.2.1",
    "notion-client>=2.7.0",
    "jupyter>=1.1.1",
    "huggingface-hub>=0.36.0",
    "iopath>=0.1.10",
    "submitit>=1.5.4",
    "torchmetrics>=1.8.2",
    "zstandard>=0.25.0",
    "tensorboard>=2.20.0",
    "fvcore>=0.1.5.post20221221",
    "lightning>=2.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black==24.2.0",
    "ufmt==2.8.0",
    "ruff-api==0.1.0",
    "usort==1.0.2",
    "gitpython==3.1.31",
    "yt-dlp",
    "opencv-python",
    "pycocotools",
    "numba",
    "python-rapidjson",
]
notebooks = [
    "notebook",
    "ipywidgets",
    "ipycanvas",
    "ipympl",
    "pycocotools",
    "decord",
    "opencv-python",
]
train = [
    "submitit",
    "tensorboard",
    "zstandard",
    "scipy",
    "torchmetrics",
    "fvcore",
    "fairscale",
]

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"

[tool.uv.sources]
torch = [
  { index = "pytorch-cu130", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu130", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
markupsafe = [
  { index = "pypi", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
